Made changes to support pickle file generation.
